requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python
keras
werkzeug==2.3.7
  
